Refector: usePreventScroll 커스텀 훅 변경
