+++ src/components/Modal/Modal.jsx
@@ -1,60 +1,62 @@
 import { motion } from "motion/react";
 import { Icon } from "@components/Icon";
 import styles from "./Modal.module.css";
-import { forwardRef, useEffect, useImperativeHandle, useRef, useState } from "react";
+import { forwardRef, useImperativeHandle, useState } from "react";
 import usePreventScroll from "./usePreventScroll";
 import { createPortal } from "react-dom";
 
 /**
- * @param {object} props
- * @param {function} props.handleToggleModal - 모달 토글 (열기/닫기)
- * @param {string} props.title - 모달 타이틀
- * @param {string} props.icon - 모달 아이콘명
- * @param {children} props.children - 모달 콘텐츠
+ * 모달 오버레이 애니메이션의 상태를 정의한 객체입니다.
+ * @type {Object}
  */
-
 const overlay = {
   hidden: { opacity: 0 },
   show: { opacity: 1 },
 };
 
+/**
+ * 모달 컨테이너 애니메이션의 상태를 정의한 객체입니다.
+ * @type {Object}
+ */
 const container = {
   hidden: { opacity: 0, scale: 0.4 },
   show: { opacity: 1, scale: 1 },
 };
 
+/**
+ * 모달 컴포넌트입니다. 모달 오버레이와 콘텐츠를 렌더링하며,
+ * 부모 컴포넌트에서 제공된 메서드를 통해 모달을 열고 닫을 수 있습니다.
+ *
+ * @component
+ * @example
+ * const modalRef = useRef(null);
+ * <Modal ref={modalRef} title="모달 제목" icon="message">모달 내용</Modal>;
+ *
+ * modalRef.current.open(); // 모달 열기
+ * modalRef.current.close(); // 모달 닫기
+ *
+ * @param {Object} props - Modal 컴포넌트의 속성입니다.
+ * @param {string} [props.title] - 모달 헤더에 표시할 제목입니다.
+ * @param {string} [props.icon] - 헤더에 표시할 아이콘의 이름입니다. 예: "info", "close" 등.
+ * @param {React.ReactNode} [props.children] - 모달의 본문에 표시할 콘텐츠입니다.
+ *
+ * @returns {React.Component|null} 열려있을 경우 렌더링되는 Modal 컴포넌트를 반환하고, 그렇지 않으면 null을 반환합니다.
+ */
 export const Modal = forwardRef(function Modal({ title, icon, children }, ref) {
+  /**
+   * 모달의 열림/닫힘 상태를 관리하는 상태입니다.
+   * @type {boolean}
+   */
   const [isOpen, setIsOpen] = useState(false);
-  const scrollPositionRef = useRef(0);
-  //const { preventScroll, allowScroll } = usePreventScroll();
 
-  useEffect(() => {
-    if (isOpen) {
-      scrollPositionRef.current = window.scrollY;
+  /**
+   * 모달이나 특정 UI 요소가 열릴 때 스크롤을 잠그는 커스텀 훅입니다.
+   */
+  usePreventScroll(isOpen);
 
-      document.body.style.position = "fixed";
-      document.body.style.width = "100%";
-      document.body.style.top = `-${scrollPositionRef.current}px`;
-      document.body.style.overflowY = "hidden";
-    } else {
-      document.body.style.position = "";
-      document.body.style.width = "";
-      document.body.style.top = "";
-      document.body.style.overflowY = "";
-      window.scrollTo(0, scrollPositionRef.current);
-    }
-
-    return () => {
-      document.body.style.position = "";
-      document.body.style.width = "";
-      document.body.style.top = "";
-      document.body.style.overflowY = "";
-      if (scrollPositionRef.current) {
-        window.scrollTo(0, scrollPositionRef.current);
-      }
-    };
-  }, [isOpen]);
-
+  /**
+   * 부모 컴포넌트에게 open과 close 메서드를 전달하기 위한 useImperativeHandle 훅 사용
+   */
   useImperativeHandle(ref, function () {
     return {
       open() {
@@ -66,6 +68,7 @@ export const Modal = forwardRef(function Modal({ title, icon, children }, ref) {
     };
   });
 
+  // 모달이 열리지 않으면 아무것도 렌더링하지 않음
   if (!isOpen) {
     return null;
   }
